Добавлены  context-файлы для управления входом/авторизацией.
